+++ models/investments.go
@@ -48,6 +48,11 @@ type InvestmentView struct {
 	Curr Currency `gorm:"-"`
 }
 
+type InvestmentWithLastPayment struct {
+	Investment
+	LastPayment time.Time
+}
+
 // GetInvestments retuns user investments
 func GetInvestments(userID uint) (res []InvestmentView, err error) {
 	sql := `select plans.id as plan_id, plans.currency, investments.amount, investments.created_at,
@@ -59,6 +64,40 @@ func GetInvestments(userID uint) (res []InvestmentView, err error) {
 
 	for i, plan := range res {
 		res[i].Curr = GetCurrency(plan.Currency)
+	}
+
+	return
+}
+
+type investmentsModel int
+
+// Investments tool models calling Investments api
+// TODO: check investment duration
+var Investments investmentsModel
+
+func (investmentsModel) NeedsProfit() (res []InvestmentWithLastPayment, err error) {
+	var (
+		// todo move this
+		profitPeriod  = 1 * time.Minute
+		needUpdatedAt = time.Now().Add(-profitPeriod)
+	)
+
+	sql := `SELECT i.id, i.created_at, i.updated_at, i.deleted_at, i.plan_id, i.user_id, i.amount, i.txn_id, p.created_at as last_payment, max(p.id) as noop
+          FROM investments i
+					LEFT JOIN profits p
+					ON i.id = p.object_id
+					WHERE (p.profit_type = ?
+					  OR p.profit_type IS NULL)
+					  AND i.deleted_at IS NULL
+					  AND (p.created_at <= ?
+							OR p.created_at IS NULL
+						)
+					GROUP BY i.id
+						`
+
+	err = db.Raw(sql, ProfitInvestment, needUpdatedAt).Scan(&res).Error
+	if err != nil {
+		return
 	}
 
 	return