+++ pkg/setting/setting.go
@@ -4,91 +4,71 @@
 
 package setting
 
-import ini "gopkg.in/ini.v1"
+import (
+	"io/ioutil"
+
+	yaml "gopkg.in/yaml.v2"
+)
 
 var (
-	confFile = "conf/app.ini"
+	confFile = "conf/app.yaml"
 
 	// Dev represent is dev mode running
 	Dev bool
 
 	// App app.ini will be mapped to this var
 	App struct {
-		SecretNumber int
-		BotName      string
-		Token        string
+		SecretNumber int `yaml:"secret_number"`
+
+		Telegram struct {
+			BotToken    string `yaml:"bot_token"`
+			BotUsername string `yaml:"bot_username"`
+		}
 
 		Web struct {
 			Port int
 		}
 
-		PaymentSystems struct {
-			QiwiTokens     []string
-			QiwiAdminToken string
+		Languages Languages
 
-			AdvcashEmail       string
-			AdvcashAPIPassword string
-			AdvcashAPIName     string
-		}
+		PaymentSystems PaymentSystems `yaml:"payment_systems"`
 
-		Account struct {
-			MinimalUSDDeposit float64
-			MinimalBTCDeposit float64
-
-			MinimalUSDReinvestAmount float64
-			MinimalBTCReinvestAmount float64
-		}
-
-		Deposit struct {
-			ReinvestAdditional float64
-
-			Lvl1Percent float64
-			Lvl2Percent float64
-			Lvl3Percent float64
-			Lvl4Percent float64
-			Lvl5Percent float64
-			Lvl6Percent float64
-
-			Lvl1USDMinSum float64
-			Lvl2USDMinSum float64
-			Lvl3USDMinSum float64
-			Lvl4USDMinSum float64
-			Lvl5USDMinSum float64
-			Lvl6USDMinSum float64
-
-			Lvl1BTCMinSum float64
-			Lvl2BTCMinSum float64
-			Lvl3BTCMinSum float64
-			Lvl4BTCMinSum float64
-			Lvl5BTCMinSum float64
-			Lvl6BTCMinSum float64
-		}
-
-		Team struct {
-			Lvl1Bonus float64
-			Lvl2Bonus float64
-			Lvl3Bonus float64
-
-			Lvl1USDMinValue float64
-			Lvl2USDMinValue float64
-			Lvl3USDMinValue float64
-
-			Lvl1BTCMinValue float64
-			Lvl2BTCMinValue float64
-			Lvl3BTCMinValue float64
-		}
-
-		Net struct {
-			Ref1 float64
-			Ref2 float64
-			Ref3 float64
-			Ref4 float64
-			Ref5 float64
-		}
+		Socials []struct {
+			Link  string
+			Image string
+			Alt   string
+		} `yaml:"social_networks"`
 	}
 
-	iniFile *ini.File
+	//iniFile *ini.File
 )
+
+type Languages []struct {
+	Name string
+	Code string
+}
+
+type PaymentSystems []struct {
+	Enabled   bool
+	Name      string
+	APISecret string `yaml:"api_secret"`
+	APIName   string `yaml:"api_name"`
+	WalletID  string `yaml:"wallet_id"`
+}
+
+func (l Languages) Names() (res []string) {
+	for _, v := range l {
+		res = append(res, v.Name)
+	}
+	return
+}
+
+func (l Languages) Codes() (res []string) {
+	for _, v := range l {
+		res = append(res, v.Code)
+	}
+	return
+}
 
 func NewContext(ops ...func()) (err error) {
 
@@ -96,11 +76,22 @@ func NewContext(ops ...func()) (err error) {
 		v()
 	}
 
-	iniFile, err = ini.Load(confFile)
+	data, err := ioutil.ReadFile(confFile)
 	if err != nil {
 		return
 	}
-	err = iniFile.MapTo(&App)
+
+	err = yaml.Unmarshal(data, &App)
+	if err != nil {
+		return
+	}
+
+	// iniFile, err = ini.Load(confFile)
+	// if err != nil {
+	// 	return
+	// }
+	// iniFile.NameMapper = mapper
+	// err = iniFile.MapTo(&App)
 
 	return
 }